--- src/main/kotlin/net/ccbluex/liquidbounce/config/types/multiChoose/MultiChooseEnumListValue.kt
@@ -1,22 +1,0 @@
-package net.ccbluex.liquidbounce.config.types.multiChoose
-
-import net.ccbluex.liquidbounce.config.types.ListValueType
-import net.ccbluex.liquidbounce.config.types.NamedChoice
-import java.util.*
-
-class MultiChooseEnumListValue<T>(
-    name: String,
-    value: EnumSet<T>,
-    choices: EnumSet<T>,
-    canBeNone: Boolean = true,
-) : MultiChooseListValue<T>(
-    name,
-    value = value,
-    choices = choices,
-    canBeNone = canBeNone,
-    listType = ListValueType.Enums,
-    autoSorting = true
-) where T : Enum<T>, T : NamedChoice {
-    override val T.elementName: String
-        get() = choiceName
-}
