--- src/main/kotlin/net/ccbluex/liquidbounce/config/types/multiChoose/MultiChooseStringListValue.kt
@@ -1,21 +1,0 @@
-package net.ccbluex.liquidbounce.config.types.multiChoose
-
-import net.ccbluex.liquidbounce.config.types.ListValueType
-import java.util.*
-
-class MultiChooseStringListValue(
-    name: String,
-    value: AbstractSet<String>,
-    choices: AbstractSet<String>,
-    canBeNone: Boolean = true,
-) : MultiChooseListValue<String>(
-    name,
-    value = value,
-    choices = choices,
-    canBeNone = canBeNone,
-    listType = ListValueType.String,
-    autoSorting = false
-) {
-    override val String.elementName: String
-        get() = this
-}
