+++ src/main/kotlin/net/ccbluex/liquidbounce/script/bindings/features/ScriptSetting.kt
@@ -19,7 +19,6 @@
 package net.ccbluex.liquidbounce.script.bindings.features
 
 import net.ccbluex.liquidbounce.config.types.*
-import net.ccbluex.liquidbounce.config.types.multiChoose.MultiChooseStringListValue
 import net.ccbluex.liquidbounce.utils.input.inputByName
 import net.minecraft.client.util.InputUtil
 import org.graalvm.polyglot.Value as PolyglotValue
